requirements.txt
python-binance==1.0.19
typer==0.9.0
rich==13.7.0
python-dotenv==1.0.1
pydantic==2.6.1
